Fix behavior on iOS 11
A combination of iOS 11 video policies and changes to the embedded
player have broken the previous behavior of background videos on the
mobile site.

Previously calling seek on an unplayed YT video would display the frame
at that timestamp. Now, calling seek on an unplayed video does nothing.
Instead, seek only works and changes the video when it's already
playing.

Because inline autoplay videos are possible on iOS 11, we can just mute
the video and autoplay it in the background to get a similar effect.